requirements.txt
openpyxl>=3.1.2
pandas>=2.0
numpy>=1.25
yfinance>=0.2.29
requests>=2.31
feedparser>=6.0.10
python-dotenv>=1.0
schedule>=1.2.0
lxml>=4.9
resend